refactoring DataStorage.js
