+++ src/core/DataStorage/ThemesStorage.js
@@ -1,0 +1,3 @@
+import { DataStorage } from "./DataStorage";
+
+export class ThemesStorage extends DataStorage {}
