+++ src/sass/_variables.sass
@@ -5,3 +5,4 @@ $padding_task_panel: 7px 7px 7px 15px
 $main_color: #2D333B
 $text_font_size: 14px
 $date_font_size: 10px
+$hover_bg: rgba(255,255,255,0.8)
