+++ src/sass/todo_main.sass
@@ -23,7 +23,7 @@
         transition-duration: 0.3s
         padding: $padding_task_panel
         &:hover
-            background: rgba(255,255,255,1)
+            background: $hover_bg
         & .task__done
             +task_button()
             +button_center()
