+++ src/sass/todo_taskPanel.sass
@@ -6,7 +6,7 @@
     transition-duration: 0.3s
     padding: $padding_task_panel
     &:hover
-        background: rgba(255,255,255,1)
+        background: $hover_bg
     & .create__btn
         min-width: 20px
         min-height: 20px
